[move][go]ディレクトリ構成の変更
