+++ define/define.go
