+++ model/scans.go
