--- server/log/check.log
@@ -1,17 +1,0 @@
-[start]2016/09/25 00:21:18 2016-09-25 00:21:18.357274262 +0900 JST
-[end]2016/09/25 00:21:22 2016-09-25 00:21:22.760559188 +0900 JST
-[start]2016/09/25 11:18:26 2016-09-25 11:18:26.281188364 +0900 JST
-[end]2016/09/25 11:18:29 2016-09-25 11:18:29.563215158 +0900 JST
-[start]2016/09/27 13:14:26 2016-09-27 13:14:26.277333464 +0900 JST
-[end]2016/09/27 13:14:50 2016-09-27 13:14:50.352361666 +0900 JST
-[start]2016/09/30 16:04:05 2016-09-30 16:04:05.196864416 +0900 JST
-[start]2016/09/30 16:05:05 2016-09-30 16:05:05.892919771 +0900 JST
-[end]2016/09/30 16:05:15 2016-09-30 16:05:15.681953362 +0900 JST
-[start]2016/09/30 16:06:20 2016-09-30 16:06:20.777728155 +0900 JST
-[end]2016/09/30 16:06:33 2016-09-30 16:06:33.848354732 +0900 JST
-[start]2016/10/09 20:30:18 2016-10-09 20:30:18.840343781 +0900 JST
-[end]2016/10/09 20:30:22 2016-10-09 20:30:22.474472416 +0900 JST
-[start]2016/10/09 20:35:13 2016-10-09 20:35:13.547532447 +0900 JST
-[end]2016/10/09 20:35:17 2016-10-09 20:35:17.36783486 +0900 JST
-[start]2016/10/09 20:35:19 2016-10-09 20:35:19.883116271 +0900 JST
-[end]2016/10/09 20:35:22 2016-10-09 20:35:22.636534092 +0900 JST
